+++ srtp/CryptoContextCtrl.h
@@ -204,12 +204,12 @@ class CryptoContextCtrl {
      * sequence number.
      *
      * @param newSeqNumber
-     *    The sequence number of the received RTP packet in host order.
+     *    The sequence number of the received RTCP packet in host order.
      *
      * @return <code>true</code> if no replay, <code>false</code> if packet
      *    is too old ar was already received.
      */
-     bool checkReplay(uint16_t newSeqNumber);
+     bool checkReplay(uint32_t newSeqNumber);
 
     /**
      * Update the SRTP packet index.
@@ -218,9 +218,9 @@ class CryptoContextCtrl {
      * 3.3.1 in the RFC when to update the ROC and ROC processing.
      *
      * @param newSeqNumber
-     *    The sequence number of the received RTP packet in host order.
+     *    The sequence number of the received RTCP packet in host order.
      */
-    void update( uint16_t newSeqNumber );
+    void update( uint32_t newSeqNumber );
 
     /**
      * Get the length of the SRTP authentication tag in bytes.
